% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_boosted_rt.R
\name{bind_boosted_rt}
\alias{bind_boosted_rt}
\title{Bind Boosting Results to Stable Rt Estimates}
\usage{
bind_boosted_rt(.data, .ref)
}
\arguments{
\item{.data}{The result of \code{summarize_rt_error()}}

\item{.ref}{Reference Rt data}
}
\value{
Full updated Rt estimates
}
\description{
\code{bind_boosted_rt()} replaces the future-conditional data points in \code{.ref}
with the updated estimates in \code{.data}.
}
\seealso{
Other internal: 
\code{\link{boost_rt}()},
\code{\link{cv_rt}()},
\code{\link{reduce_rt_error}()},
\code{\link{summarize_rt_error}()},
\code{\link{validate_rt}()},
\code{\link{weight_rt_error}()}
}
\concept{internal}
