% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_boosted_rt.R
\name{boost_rt}
\alias{boost_rt}
\title{Boost Rt Estimates Given Reference Data and Cross-Validation Errors}
\usage{
boost_rt(.data, .ref)
}
\arguments{
\item{.data}{Results of cross-validation}

\item{.ref}{Full estimates of Rt}
}
\value{
Rt estimates corrected for smoothing uncertainty
}
\description{
\code{boost_rt()} computes estimates of a Gamma-distributed reproduction number
given the results of rolling cross-validation of the decomposition and
the reference data.
}
\seealso{
Other internal: 
\code{\link{bind_boosted_rt}()},
\code{\link{cv_rt}()},
\code{\link{reduce_rt_error}()},
\code{\link{summarize_rt_error}()},
\code{\link{validate_rt}()},
\code{\link{weight_rt_error}()}
}
\concept{internal}
