% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts_season.R
\name{bsts_season.dynamic}
\alias{bsts_season.dynamic}
\title{Unfinished}
\usage{
bsts_season.dynamic(
  state = list(),
  .data = state[[".data"]],
  method = "dynamic",
  period = NULL,
  season = NULL,
  ar = TRUE,
  sigma.prior = NULL,
  na.action = na.exclude,
  ...
)
}
\description{
Unfinished
}
