% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts_season.R
\name{bsts_season.harmonic}
\alias{bsts_season.harmonic}
\title{Method dispatch not working}
\usage{
bsts_season.harmonic(
  state = list(),
  .data = state[[".data"]],
  method = "harmonic",
  period = NULL,
  season = NULL,
  sigma.prior = NULL,
  initial.state.prior = NULL,
  sdy = NULL
)
}
\description{
Method dispatch not working
}
