% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts_season.R
\name{bsts_season}
\alias{bsts_season}
\title{Bayesian Structural Time Series: Seasonal Components}
\usage{
bsts_season(
  state = list(),
  .data = state[[".data"]],
  period = NULL,
  season = NULL,
  ...
)
}
\arguments{
\item{state}{A list of state components you wish to add to. If omitted,
an empty list will be assumed. This argument is named \code{state.specification}
in bsts.}

\item{period}{The length of a full seasonal cycle. Either a time-based definition (e.g. "1 week")
or the number of observations in a season (e.g. 7). Supplying a
number makes this equivalent to \code{nseasons} in
\code{\link[bsts:AddSeasonal]{AddSeasonal()}} or \code{period} in
\code{\link[bsts:AddTrig]{AddTrig()}} (for methods \code{regression} or
\code{harmonic}, respectively).}

\item{season}{The length of a season within a period (e.g. "1 month" for a 1 year period
with monthly seasonality); the default is one season per
observation within a period. Can be supplied in the same
way as \code{period}. This parameter is ignored for
\code{method = "harmonic"}.}

\item{...}{Additional arguments to pass to methods; see the methods above
for details}

\item{method}{Which seasonal model to use. Choose from \code{"regression"} or
\code{"harmonic"}.}
}
\description{
\code{bsts_season()} is a generic that wraps the two seasonal models in the
bsts package into a user-friendly interface. Model-specific arguments are
passed via \code{...}; see the two methods (
\code{\link[bsts_seasonal.regression]{"regression"}} and
\code{\link[bsts_seasonal.harmonic]{"harmonic"}}
) for details of those arguments. \code{bsts_seasonal()} also uses the timetk
package to make model specification easier or, if you prefer, automate it
entirely based on tunable heuristics.
}
