% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts_trend.R
\name{bsts_trend.level}
\alias{bsts_trend.level}
\title{Local Level Trend State Component}
\usage{
\method{bsts_trend}{level}(
  state = list(),
  .data = state[[".data"]],
  method = "level",
  sigma.prior = NULL,
  initial.state.prior = NULL,
  sdy = NULL,
  initial.y = NULL
)
}
\description{
Local Level Trend State Component
}
