% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts_trend.R
\name{bsts_trend.local}
\alias{bsts_trend.local}
\title{Local Linear Trend State Component}
\usage{
\method{bsts_trend}{local}(
  state = list(),
  .data = state[[".data"]],
  method = "local",
  level.sigma.prior = NULL,
  slope.sigma.prior = NULL,
  initial.level.prior = NULL,
  initial.slope.prior = NULL,
  sdy = NULL,
  initial.y = NULL
)
}
\description{
Local Linear Trend State Component
}
