% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts_trend.R
\name{bsts_trend}
\alias{bsts_trend}
\alias{bsts_trend.semilocal}
\alias{bsts_trend.local}
\alias{bsts_trend.robust}
\alias{bsts_trend.level}
\title{Bayesian Structural Time Series: Trend Components}
\usage{
bsts_trend(
  state = list(),
  .data = state[[".data"]],
  method = c("semilocal", "local", "robust", "level"),
  ...
)

\method{bsts_trend}{semilocal}(
  state = list(),
  .data = state[[".data"]],
  method = "semilocal",
  level.sigma.prior = NULL,
  slope.mean.prior = NULL,
  slope.ar1.prior = NULL,
  slope.sigma.prior = NULL,
  initial.level.prior = NULL,
  initial.slope.prior = NULL,
  sdy = NULL,
  initial.y = NULL
)

\method{bsts_trend}{local}(
  state = list(),
  .data = state[[".data"]],
  method = "local",
  level.sigma.prior = NULL,
  slope.sigma.prior = NULL,
  initial.level.prior = NULL,
  initial.slope.prior = NULL,
  sdy = NULL,
  initial.y = NULL
)

\method{bsts_trend}{robust}(
  state = list(),
  .data = state[[".data"]],
  method = "robust",
  save.weights = FALSE,
  level.sigma.prior = NULL,
  level.nu.prior = NULL,
  slope.sigma.prior = NULL,
  slope.nu.prior = NULL,
  initial.level.prior = NULL,
  initial.slope.prior = NULL,
  sdy = NULL,
  initial.y = NULL
)

\method{bsts_trend}{level}(
  state = list(),
  .data = state[[".data"]],
  method = "level",
  sigma.prior = NULL,
  initial.state.prior = NULL,
  sdy = NULL,
  initial.y = NULL
)
}
\arguments{
\item{state}{A list of state components you wish to add to. If omitted,
an empty list will be assumed. This argument is named \code{state.specification}
in bsts.}

\item{.data}{The time series to be modeled, as a numeric vector. Unlike bsts,
this is piped forward as part of the state if defined, so you only need
to specify it once (at the beginning of the model-building pipeline).}

\item{method}{Which trend model to use. Choose from \code{"semilocal"}
(the default), \code{"local"}, \code{"robust"}, or \code{"level"}.}

\item{...}{Additional arguments to pass to methods; see the methods above
for details}

\item{level.sigma.prior}{An object created by
   \code{\link[Boom]{SdPrior}} describing the prior
   distribution for the standard deviation of the level component.}

\item{slope.mean.prior}{An object created by
   \code{\link[Boom]{NormalPrior}} giving the prior distribution for
   the mean parameter in the generalized local linear trend model (see
   below). }

\item{slope.ar1.prior}{An object created by
   \code{\link[Boom]{Ar1CoefficientPrior}} giving the prior
   distribution for the ar1 coefficient parameter in the generalized
   local linear trend model (see below). }

\item{slope.sigma.prior}{An object created by
   \code{\link[Boom]{SdPrior}} describing the prior distribution of
   the standard deviation of the slope component.}

\item{initial.level.prior}{An object created by
    \code{\link[Boom]{NormalPrior}} describing the initial distribution
    of the level portion of the initial state vector.}

\item{initial.slope.prior}{An object created by
    \code{\link[Boom]{NormalPrior}} describing the prior distribution
    for the slope portion of the initial state vector.}

\item{sdy}{The standard deviation of the series to be modeled.  This
    will be ignored if \code{y} is provided, or if all the required
    prior distributions are supplied directly.  }

\item{initial.y}{The initial value of the series being modeled.  This will be
    ignored if \code{y} is provided, or if the priors for the initial
    state are all provided directly.}

\item{save.weights}{A logical value indicating whether to save the
    draws of the weights from the normal mixture representation.}

\item{level.nu.prior}{An object inheritng from the class
    \code{\link[Boom]{DoubleModel}}, representing the prior
    distribution on the \code{nu} tail thickness parameter of the T
    distribution for errors in the evolution equation for the level
    component.}

\item{slope.nu.prior}{An object inheritng from the class
    \code{\link[Boom]{DoubleModel}}, representing the prior
    distribution on the \code{nu} tail thickness parameter of the T
    distribution for errors in the evolution equation for the slope
    component.}

\item{sigma.prior}{An object created by \code{\link[Boom]{SdPrior}}
    describing the prior distribution for the standard deviation of the
    random walk increments.}

\item{initial.state.prior}{An object created using
    \code{\link[Boom]{NormalPrior}}, describing the prior distribution
    of the initial state vector (at time 1).}
}
\value{
A list with the elements necessary to specify the chosen trend model
}
\description{
\code{bsts_trend()} is a generic that wraps the various trend models in the
bsts package into a user-friendly interface. Model-specific arguments are
passed via \code{...}; see the four methods for details of those arguments.
}
\seealso{
\code{\link[bsts:bsts]{bsts()}},
\code{\link[bsts:AddSemilocalLinearTrend]{AddSemilocalLinearTrend()}},
\code{\link[bsts:AddLocalLinearTrend]{AddLocalLinearTrend()}},
\code{\link[bsts:AddStudentLocalLinearTrend]{AddStudentLocalLinearTrend()}},
\code{\link[bsts:AddLocalLevel]{AddLocalLevel()}}

Other bsts: 
\code{\link{bsts_season}()}
}
\concept{bsts}
