% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts_trend.R
\name{bsts_trend.robust}
\alias{bsts_trend.robust}
\title{Robust Local Linear Trend State Component}
\usage{
\method{bsts_trend}{robust}(
  state = list(),
  .data = state[[".data"]],
  method = "robust",
  save.weights = FALSE,
  level.sigma.prior = NULL,
  level.nu.prior = NULL,
  slope.sigma.prior = NULL,
  slope.nu.prior = NULL,
  initial.level.prior = NULL,
  initial.slope.prior = NULL,
  sdy = NULL,
  initial.y = NULL
)
}
\description{
Robust Local Linear Trend State Component
}
