% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts_trend.R
\name{bsts_trend.semilocal}
\alias{bsts_trend.semilocal}
\title{Semilocal Linear Trend State Component}
\usage{
\method{bsts_trend}{semilocal}(
  state = list(),
  .data = state[[".data"]],
  method = "semilocal",
  level.sigma.prior = NULL,
  slope.mean.prior = NULL,
  slope.ar1.prior = NULL,
  slope.sigma.prior = NULL,
  initial.level.prior = NULL,
  initial.slope.prior = NULL,
  sdy = NULL,
  initial.y = NULL
)
}
\description{
Semilocal Linear Trend State Component
}
