% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_active_cases.R
\name{calc_active_cases}
\alias{calc_active_cases}
\title{Calculate Active Cases Based on 14 Day Heuristic}
\usage{
calc_active_cases(
  data = coviData::process_positive_people(date = date),
  date = NULL,
  quiet = FALSE
)
}
\arguments{
\item{data}{Case data from the investigation file, as output by
\code{\link[coviData:process_positive_people]{process_positive_people()}}}

\item{date}{The date of the data file to use; if \code{NULL}, the most recent data
available will be used}

\item{quiet}{Should info messages be silenced?}
}
\value{
An \code{integer} number of active cases on \code{date}
}
\description{
Calculate Active Cases Based on 14 Day Heuristic
}
