% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_linelist.R
\name{clean_linelist}
\alias{clean_linelist}
\title{Remove Incomplete, Illogical, and Missing Data from an Incidence Linelist}
\usage{
clean_linelist(
  .data,
  .collection_date = "collection_date",
  .report_date = "report_date",
  start_date = "2020-03-12",
  delay_period = 14L,
  pct_reported = 0.9
)
}
\arguments{
\item{.data}{A data frame containing one incident observation per row}

\item{.collection_date}{\verb{<tidy-select>} A \code{Date} column to use as the
collection date of the observed case}

\item{.report_date}{\verb{<tidy-select>} A \code{Date} column to use as the report
date of the observed case}

\item{start_date}{The start date of the epidemic;
defaults to \code{"2020-03-12"}, which is the beginning of the contiguous
part of Shelby County's observed cases (at least one case observed per
day since that date).}

\item{delay_period}{The length of time to use in calculating reporting
delay; can be a time-based definition (e.g. "2 weeks") or an integer number
of days. If \code{NULL}, \code{delay_period} is set to \code{"14 days"}.}

\item{pct_reported}{The percent of total cases reported before considering
a collection date to be fully observed. It is not recommended to set this
to \code{1}, as reporting delays typically contain very large outliers which
will skew the results. The default is \code{0.9}, which strikes a balance
between sensitivity and robustness in Shelby County data.}
}
\value{
A \code{tibble} containing the cleaned linelist
}
\description{
\code{clean_linelist()} creates a cleaned version of an incidence linelist with
collection and report dates. It removes observations with collection dates
that have not been fully reported (as determined by \code{pct_reported}), as well
as observations with missing collection dates or collection dates prior to
\code{start_date}. It then checks that all report dates are on or after the
corresponding collection dates and removes observations where this is untrue.
}
\seealso{
Higher-level functions
\code{
  \link[covidModel:prep_linelist_decomposition]{prep_linelist_decomposition()}
  }, \code{\link[covidModel:prep_linelist]{prep_linelist()}}
}
