% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidModel-package.R
\docType{package}
\name{covidModel-package}
\alias{covidModel}
\alias{covidModel-package}
\title{covidModel: COVID-19 Modeling Tools for the Shelby County Health Department}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

covidModel contains inference and prediction
    functions to aid understanding of the COVID-19 epidemic in Shelby
    County. The focus is primarily on temporal data, though other methods
    may be included.
}
\author{
\strong{Maintainer}: Jesse Smith \email{jesse.smith@shelbycountytn.gov} (\href{https://orcid.org/0000-0001-7375-2463}{ORCID})

}
\keyword{internal}
