% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_decomposition.R
\name{cv_decomposition}
\alias{cv_decomposition}
\title{Cross-Validate Predictions in an STL Decomposition Conditional on Future Data}
\usage{
cv_decomposition(
  .data,
  .col = "observed_cleaned",
  trend = "31 days",
  period = "7 days"
)
}
\arguments{
\item{.data}{A data frame containing the time series data to resample}

\item{.col}{The column containing the data to resample}

\item{trend}{The length of time to use in trend decomposition; can be a
time-based definition (e.g. "1 month") or an integer number of days. If
\code{NULL} or \code{"auto"}, \code{trend} is set automatically using the tunable
heuristics in the timetk package.}

\item{period}{The length of time to use in seasonal decomposition; can be a
time-based definition (e.g. "1 week") or an integer number of days. If
\code{NULL} or \code{"auto"}, \code{period} is set automatically using the tunable
heuristics in the timetk package.}
}
\value{
A list of \code{tibble} objects, each containing the results of one
sampling step for the dates in \code{start_date + trend} to
\code{end_date - trend/2}, where \code{end_date} is the last completely observed
date. See the \code{Value} section of
\code{\link[covidModel:validate_decomposition]{validate_decomposition()}}
for information on the components of each sample.
}
\description{
\code{cv_decomposition()} applies rolling cross-validation to the portion of an
STL decomposition conditional on future data. It provides
the predicted values and associated "errors" for each "forecast" horizon and
decomposed component across all stably-estimable time points.
}
\details{
When smoothing time series using LOESS, a portion of the smooth will change
as future data comes in (approximately the last \code{trend/2} data points). These
points are said to be conditional on future data, and the difference between
the smooth with and without future data can be computed for all time points
that have already reached the "stable" portion of the series. This is done by
rolling the decomposition step across the time series and comparing the
smoothed point at each "forecast" horizon to the stable estimates. In effect,
this is rolling cross-validation of the STL decomposition using the fully
smoothed data as the reference values.
}
\seealso{
The workhorse function
\code{\link[covidModel:validate_decomposition]{validate_decomposition()}}
and the higher-level function
\code{
  \link[covidModel:cv_linelist_decomposition]{cv_linelist_decomposition()}
  }
}
\keyword{internal}
