% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_boosted_rt.R
\name{cv_rt}
\alias{cv_rt}
\title{Apply Rolling Cross-Validation to Rt Calculations Given Cross-Validated
Decomposition Estimates}
\usage{
cv_rt(.data, .ref, .t, serial_interval_mean, serial_interval_sd, half_trend)
}
\arguments{
\item{.data}{The cross-validation results from \code{cv_linelist_decomposition()}}

\item{.ref}{Reference Rt}

\item{.t}{The quoted name of a date column corresponding to the observations
in \code{incid}}

\item{serial_interval_mean}{The average number of days between infection of a
primary case and a secondary case}

\item{serial_interval_sd}{The standard deviation of the number of days
between infection of a primary case and a secondary case}

\item{half_trend}{The portion of the trend conditional on future data}
}
\value{
A list of tibbles with Rt estimates and associated errors
}
\description{
\code{cv_rt()} computes the reproduction number for each sample in the results
of \code{
\link[covidModel:cv_linelist_decomposition]{cv_linelist_decomposition()}
}.
}
\seealso{
Other internal: 
\code{\link{bind_boosted_rt}()},
\code{\link{boost_rt}()},
\code{\link{reduce_rt_error}()},
\code{\link{summarize_rt_error}()},
\code{\link{validate_rt}()},
\code{\link{weight_rt_error}()}
}
\concept{internal}
