% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deanomalize.R
\name{deanomalize}
\alias{deanomalize}
\title{Replace Anomalies with Expected Values Using the \code{anomalize} Package}
\usage{
deanomalize(
  .data,
  .col,
  period = NULL,
  trend = NULL,
  cutoff = 0.05,
  quiet = FALSE,
  plot = FALSE,
  ...
)
}
\arguments{
\item{.data}{A data frame containg a time-based column and a column of
observations. The time-based column will be identified automatically. Can
include other columns, but these will be dropped.}

\item{.col}{The column containing the observations at each time point}

\item{period}{The length of time to use in seasonal decomposition; can be a
time-based definition (e.g. "1 week") or an integer number of days. If
\code{NULL} or \code{"auto"}, \code{period} is set automatically using the tunable
heuristics in the timetk package.}

\item{trend}{The length of time to use in trend decomposition; can be a
time-based definition (e.g. "1 month") or an integer number of days. If
\code{NULL} or \code{"auto"}, \code{trend} is set automatically using the tunable
heuristics in the timetk package.}

\item{cutoff}{The cutoff value for anomaly detection; controls both the
maximum percentage of data points that may be considered anomalies, as well
as the critical value for the Generalized Extreme Studentized Deviate test
used to detect the anomalies. Can be interpreted as the desired maximum
probability that an individual data point is labeled an anomaly.}

\item{quiet}{Should messages and warnings be suppressed?}

\item{plot}{Should anomalies be plotted for visual inspection?}

\item{...}{Additional arguments to pass to
\code{\link[anomalize:time_decompose]{time_decompose(method = "stl")}}}
}
\value{
A \code{tibble} with a column for the time index and columns \code{observed},
\code{season}, \code{trend}, \code{remainder}, \code{remainder_l1}, \code{remainder_l2}, \code{anomaly},
and \code{observed_cleaned}
}
\description{
\code{deanomalize()} wraps the workflow from the anomalize package into a single
function. This includes decomposition into seasonal, trend, and remainder
components using robust STL; identification of anomalies in the remainders
using an iterative Generalize Extreme Studentized Deviate Test; and
replacment of the identified anomalies with the seasonal + trend expected
value.
}
