% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_activity.R
\name{estimate_activity}
\alias{estimate_activity}
\alias{estimate_activity.data.frame}
\alias{estimate_activity.covidmodel_rt}
\title{Estimate Infectious Activity from Incidence and Serial Interval}
\usage{
estimate_activity(.data, incid, .t, serial_interval_mean, serial_interval_sd)

\method{estimate_activity}{data.frame}(
  .data,
  incid = "trend",
  .t = "collection_date",
  serial_interval_mean = 6,
  serial_interval_sd = 4.17
)

\method{estimate_activity}{covidmodel_rt}(
  .data,
  incid = ".incid",
  .t = ".t",
  serial_interval_mean = attr(.data, "serial_interval")[[1L]],
  serial_interval_sd = attr(.data, "serial_interval")[[2L]]
)
}
\description{
Estimate Infectious Activity from Incidence and Serial Interval
}
