% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rt.R
\name{estimate_rt}
\alias{estimate_rt}
\title{Simple Bayesian Estimates of the Time-Varying Reproduction Number}
\usage{
estimate_rt(
  .data,
  incid = "trend",
  t = "collection_date",
  serial_interval_mean = 6,
  serial_interval_sd = 4.17
)
}
\arguments{
\item{.data}{A data frame containing the incidence curve and dates}

\item{incid}{The quoted name of a numeric column containing the incidence
curve}

\item{t}{The quoted name of a date column corresponding to the observations
in \code{incid}}

\item{serial_interval_mean}{The average number of days between infection of a
primary case and a secondary case}

\item{serial_interval_sd}{The standard deviation of the number of days
between infection of a primary case and a secondary case}
}
\value{
A \code{tibble} with columns \code{.t}, \code{.pred} (the median), \code{.pred_lower}
(the lower bound of the 95\% credible interval), \code{.pred_upper}
(the upper bound of the 95\% credible interval), \code{.mean} (the average), and
\code{.cv} (the coefficient of variation)
}
\description{
\code{estimate_rt()} uses the methodology of Cori et al 2013 to estimate the time-
varying reproduction number. It calls
\code{\link[EpiEstim:estimate_R]{estimate_R(method = "parametric_si")}} under
the hood. It purposefully does not support aggregating counts over multiple
time periods due to poor coverage of the resulting credible intervals, as
well as shifting of the results due to lagging of the results. Use
\code{smooth_for_rt()} to pre-smooth the data instead.
}
