% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rt.R
\name{expm1_decomposed}
\alias{expm1_decomposed}
\title{Convert Decompositions Performed Under a \code{log1p()} Transform to Linear Scale}
\usage{
expm1_decomposed(.data)
}
\arguments{
\item{.data}{The output of \code{time_decompose()}}
}
\value{
The input decomposition with \code{expm1()} applied appropriately
}
\description{
\code{expm1_decomposed()} converts the output of
\code{\link[anomalize:time_decompose]{time_decompose()}} back to the linear
scale when the input to \code{time_decompose()} is on the log scale. This amounts
to selective sequential application of \code{\link[base:expm1]{expm1()}} to
the component columns.
}
\details{
Because log-scale additive decomposition is multiplicative decomposition on
the linear scale, functions which automatically perform decomposition on the
log scale yield unexpected output. To remedy this, an approximate conversion
back to the linear scale can be performed. The steps are as follows (with all
exponentiation follow by an "add-1" operation):

\enumerate{
\item \strong{observed}: Exponentiate the \code{observed} column
\item \strong{trend}: Exponentiate the \code{trend} column
\item \strong{season}: Exponentiate \code{trend + season} and subtract the
exponentiated \code{trend}
\item \strong{remainder} Subtract the exponentiated \code{trend} and \code{season}
from the exponentiated \code{observed} values (\code{observed - trend - season})
}

The underlying code is perhaps a simpler explanation than the above; consult
it for further exposition.

Exponentiating the \code{observed} values is a simple inverse transform; however,
the component columns require some assumptions to convert back to a linear
scale. This process uses the \code{trend} as a "base" and simply exponentiates it
as in \code{observed}. This can then be used to derive the \code{season} component, and
the \code{remainder} is whatever is left over. While this is not a theoretically
valid process (there is none for this operation), it is consistent with
treating the \code{trend} as a smoothed version of \code{observed}, which is usually
done in practice. Once this assumption is made, the \code{season} transformation
becomes valid. The \code{remainder} is then assumed to be the difference between
the observed values (\code{observed}) expected values (\code{season + trend}), as
usual.
}
\seealso{
\code{\link[anomalize:time_decompose]{time_decompose()}},
\code{\link[base:expm1]{expm1()}}
}
\keyword{internal}
