% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{expmC}
\alias{expmC}
\title{"Exp(x) - C" Transform}
\usage{
expmC(x, C = exp(1))
}
\arguments{
\item{x}{A numeric vector to be transformed}

\item{C}{A constant; defaults to square root of machine precision}
}
\description{
"Exp(x) - C" Transform
}
