% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{functionalize_vec}
\alias{functionalize_vec}
\title{Convert a Time-Indexed Vector to Function of Time (i.e. x = f(t))}
\usage{
functionalize_vec(x, t = NULL)
}
\description{
Convert a Time-Indexed Vector to Function of Time (i.e. x = f(t))
}
