% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rt.R
\name{gamma-distr-conversion}
\alias{gamma-distr-conversion}
\alias{gamma_shape}
\alias{gamma_rate}
\alias{gamma_mean}
\alias{gamma_sd}
\title{Convert Between Mean/SD and Shape/Rate of Gamma Distribution}
\usage{
gamma_shape(mean, sd)

gamma_rate(mean, sd)

gamma_mean(shape, rate)

gamma_sd(shape, rate)
}
\arguments{
\item{mean}{A vector of one or more means}

\item{sd}{A vector of one or more standard deviations}

\item{shape}{A vector of one or more shape parameters}

\item{rate}{A vector of one or more rate parameters}
}
\description{
These functions return the mean, standard deviation, shape, or rate of a
gamma distribution given the other pair of parameters.
}
\keyword{internal}
