% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate a Continuous Function of Time
\code{interpolate()} creates a function of time from a time-indexed input object.
It powers \code{functionalize()}.}
\usage{
interpolate(.data, .x, .t = NULL, rtn_data = FALSE)
}
\arguments{
\item{.data}{A input dataset with a column to interpolate}
}
\description{
Interpolate a Continuous Function of Time
\code{interpolate()} creates a function of time from a time-indexed input object.
It powers \code{functionalize()}.
}
