% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{logCp}
\alias{logCp}
\title{"Log(x + C)" Transform}
\usage{
logCp(x, C = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{A numeric vector to be transformed}

\item{C}{A constant; defaults to square root of machine precision}
}
\description{
"Log(x + C)" Transform
}
