% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast_cases}
\alias{nowcast_cases}
\title{Nowcast Cases by Specimen Collection Date}
\usage{
nowcast_cases(
  .data,
  .collection_date = "collection_date",
  .report_date = "report_date",
  today = Sys.Date(),
  iterations = ceiling(0.5 * 10000/(1 - burn)),
  burn = 0.2
)
}
\arguments{
\item{.data}{A linelist of incidence cases as a data frame}

\item{.collection_date}{A \code{Date} column containing the dates that cases
were tested}

\item{.report_date}{A \code{Date} column containing the dates that cases were
reported}

\item{today}{The date to consider "today"}

\item{iterations}{The number of MCMC iterations for each model in the
ensemble; the total number of iterations will be twice this amount.}

\item{burn}{The percentage of initial iterations to discard for each model}
}
\description{
\code{nowcast_cases()} predicts the number of cases expected through the current
date given a linelist of cases with collection and report dates. It uses
\code{\link[covidModel:estimate_delay]{estimate_delay()}} to determine the
last complete collection date and forecasts from that point. Cases are
predicted by a weighted ensemble of two log-linear structural time series
models with semilocal and robust (Student-t) trends.
}
