% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R, R/nowcast.R
\name{nowcast_tests}
\alias{nowcast_tests}
\title{Nowcast}
\usage{
nowcast_tests(
  .data,
  .collection_date = "collection_date",
  .report_date = "report_date",
  today = Sys.Date(),
  iterations = ceiling(10000/0.8),
  burn = 0.2
)

nowcast_tests(
  .data,
  .collection_date = "collection_date",
  .report_date = "report_date",
  today = Sys.Date(),
  iterations = ceiling(10000/0.8),
  burn = 0.2
)
}
\arguments{
\item{.data}{A linelist of incident cases/tests as a data frame}

\item{.collection_date}{A \code{Date} column containing the test date}

\item{.report_date}{A \code{Date} column containing the report date}

\item{today}{The date to consider "today"}

\item{iterations}{The number of MCMC iterations for each model in the
ensemble; the total number of iterations will be twice this amount.}

\item{burn}{The percentage of initial iterations to discard for each model}
}
\description{
\code{nowcast_tests()} predicts the expected number of tests per day through the
current date given a linelist of cases with collection and report dates. It
uses \code{\link[covidModel:estimate_delay]{estimate_delay()}} to determine
the last complete collection date and forecasts from that point. Tests are
predicted by a weighted ensemble of two structural time series models with
semilocal and robust (Student-t) trends and a square-root link.
}
