% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rt}
\alias{plot_rt}
\title{Plot the Results of \code{estimate_rt()}}
\usage{
plot_rt(.data, start = "2020-04-01", end = Sys.Date(), .rough_rt = NULL)
}
\arguments{
\item{.data}{Rt estimates resulting from a call to
\code{\link[covidModel:estimate_rt]{estimate_rt()}}}

\item{start}{The start date of the x axis}

\item{end}{The end date of the x axis}

\item{.rough_rt}{Unsmoothed Rt estimates, e.g. as produced by
\code{estimate_rt(trend = 1L, boost = FALSE)}}
}
\value{
A \code{ggplot} object
}
\description{
\code{plot_rt()} plots the effective reproductive number over the course of the
epidemic in Shelby County, along with reference dates and severity levels
for comparison. Optionally, supplied unsmoothed Rt estimates to \code{.rough_rt}
will add points for this estimates to the background for reference.
}
