% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_unboosted_rt.R
\name{prep_data_rt}
\alias{prep_data_rt}
\title{Prepare Data for Rt Calculation}
\usage{
prep_data_rt(.data, .incid = NULL, .t = NULL)
}
\arguments{
\item{.data}{A data frame containing the incidence curve and dates}

\item{.incid}{The quoted name of a numeric column containing the incidence
curve}

\item{.t}{The quoted name of a date column corresponding to the observations
in \code{incid}}

\item{incid}{The quoted name of a numeric column containing the incidence
curve}

\item{t}{The quoted name of a date column corresponding to the observations
in \code{incid}}
}
\description{
\code{prep_data_rt()} is an internal function used to prepare the \code{incid} argument
for \code{\link[EpiEstim:estimate_R]{estimate_R()}}.
}
\keyword{internal}
