% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_linelist.R
\name{prep_linelist_decomposition}
\alias{prep_linelist_decomposition}
\title{Prepare Linelist for STL Decomposition of Incidence Curve}
\usage{
prep_linelist_decomposition(
  .data,
  .collection_date = "collection_date",
  .report_date = "report_date",
  start_date = "2020-03-12",
  trend = "30 days",
  period = "7 days",
  delay_period = "14 days",
  pct_reported = 0.9,
  cutoff = 0.05,
  plot_anomalies = FALSE
)
}
\arguments{
\item{.data}{A data frame containing one incident observation per row}

\item{.collection_date}{\verb{<tidy-select>} A \code{Date} column to use as the
collection date of the observed case}

\item{.report_date}{\verb{<tidy-select>} A \code{Date} column to use as the report
date of the observed case}

\item{start_date}{The start date of the epidemic;
defaults to \code{"2020-03-12"}, which is the beginning of the contiguous
part of Shelby County's observed cases (at least one case observed per
day since that date).}

\item{trend}{The length of time to use in trend decomposition; can be a
time-based definition (e.g. "1 month") or an integer number of days. The
default is \code{"30 days"}; however, if \code{NULL} or \code{"auto"}, \code{trend} is set
automatically using the tunable heuristics in the timetk package.}

\item{period}{The length of time to use in seasonal decomposition; can be a
time-based definition (e.g. "1 week") or an integer number of days. The
default is \code{"7 days"}; however, if \code{NULL} or \code{"auto"}, \code{period} is set
automatically using the tunable heuristics in the timetk package.}

\item{delay_period}{The length of time to use in calculating reporting
delay; can be a time-based definition (e.g. "2 weeks") or an integer number
of days. If \code{NULL}, \code{delay_period} is set to \code{"14 days"}.}

\item{pct_reported}{The percent of total cases reported before considering
a collection date to be fully observed. It is not recommended to set this
to \code{1}, as reporting delays typically contain very large outliers which
will skew the results. The default is \code{0.9}, which strikes a balance
between sensitivity and robustness in Shelby County data.}

\item{cutoff}{The cutoff value for outlier detection; controls both the
maximum percentage of data points that may be considered outliers, as well
as the critical value for the Generalized Extreme Studentized Deviate test
used to detect the outliers. Can be interpreted as the desired maximum
likelihood that an individual data point is labelled an outlier.}

\item{plot_anomalies}{Should anomalies be plotted for visual inspection? If
\code{TRUE}, the plot will be on the log-scale.}
}
\value{
A \code{tibble} containing cleaned and deanomalized count data on the
log-plus-1 scale, as well as a decomposition of that data and information
associated with anomaly detection
}
\description{
\code{prep_linelist_decomposition()} prepares linelist data for decomposition
by \code{\link[anomalize:time_decompose]{time_decompose(method = "stl)}}. It
is the first step in \code{\link[covidModel:prep_linelist]{prep_linelist()}}.
}
\seealso{
Component functions
\code{\link[covidModel:clean_linelist]{clean_linelist()}} and
\code{\link[covidModel:deanomalize]{deanomalize()}}, as well as
higher-level function
\code{\link[covidModel:prep_linelist]{prep_linelist()}}
}
\keyword{internal}
