% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_boosted_rt.R
\name{reduce_rt_error}
\alias{reduce_rt_error}
\title{Combine a List of Rt Errors into a Tibble Grouped by Forecast Horizon}
\usage{
reduce_rt_error(.data)
}
\arguments{
\item{.data}{A list of tibbles containing error estimates for each time point
in a time series}
}
\value{
A \code{tibble} containing error estimates by forecast horizon for each
estimated time point
}
\description{
\code{reduce_rt_error()} combine a list of tibbles containing the mapped error
estimates using \code{\link[covidModel:validate_rt]{validate_rt()}}. Reduce
here is used in the context of "Map-Reduce"; it has nothing to do with
"reducing" the size of errors. This function is used internally by
\code{\link[covidModel:estimate_boosted_rt]{estimate_boosted_rt()}}.
}
\seealso{
Other internal: 
\code{\link{bind_boosted_rt}()},
\code{\link{boost_rt}()},
\code{\link{cv_rt}()},
\code{\link{summarize_rt_error}()},
\code{\link{validate_rt}()},
\code{\link{weight_rt_error}()}
}
\concept{internal}
