% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_decomposition.R
\name{resample_decomposition}
\alias{resample_decomposition}
\title{Resample Endpoint Errors in an STL Decomposition}
\usage{
resample_decomposition(
  .data,
  .col = "observed_cleaned",
  trend = "31 days",
  period = "7 days"
)
}
\arguments{
\item{.data}{A data frame containing the time series data to resample}

\item{.col}{The column containing the data to resample}

\item{trend}{The length of time to use in trend decomposition; can be a
time-based definition (e.g. "1 month") or an integer number of days. The
default is \code{"30 days"}; however, if \code{NULL} or \code{"auto"}, \code{trend} is set
automatically using the tunable heuristics in the timetk package.}

\item{period}{The length of time to use in seasonal decomposition; can be a
time-based definition (e.g. "1 week") or an integer number of days. The
default is \code{"7 days"}; however, if \code{NULL} or \code{"auto"}, \code{period} is set
automatically using the tunable heuristics in the timetk package.}
}
\value{
A list of \code{tibble} objects, each containing the results of one
sampling step. See the \code{Value} section of
\code{\link[covidModel:sample_decomposition]{sample_decomposition()}} for
information on the components of each sample.
}
\description{
\code{resample_decomposition()} computes the right-hand endpoint errors of an
STL decomposition across all past data. Since smoothing generally depends on
both past and future observations, recent points in a smoothing algorithm
are subject to change conditional on future data (up to the limit of the
smoothing window, if one exists). \code{resample_decomposition()} provides the
previous changes for each "forecast" horizon, as well as the associated
error, for each component of the decomposition.
}
\seealso{
The workhorse function
\code{\link[covidModel:sample_decomposition]{sample_decomposition()}} and
the higher-level function
\code{
  \link[covidModel:resample_linelist_decomposition]{
  resample_linelist_decomposition()
  }}
}
\keyword{internal}
