% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma-distr.R
\name{rgamma_vec}
\alias{rgamma_vec}
\title{Vectorized \code{rgamma()} Function}
\usage{
rgamma_vec(n, shape, rate = 1)
}
\arguments{
\item{n}{Number of observations. If \code{vec_size(n) > 1L}, the size is
taken to be the number required.}

\item{shape, rate}{Shape and rate parameters for the gamma distribution.
Both must be positive, \code{scale} strictly. Inputs of size 1 will be
recycled, otherwise the sizes of \code{shape} and \code{scale} must match.}
}
\value{
A \code{tibble} of random deviates with \code{n} rows and columns corresponding
to the values in the vectors supplied to \code{shape}, \code{rate} (named with the
pattern \verb{gamma\{i\}}, where \code{i} is the \code{i}th value in the recycled \code{shape}
and \code{rate} vectors, padded by zeros to ensure consistent length).
}
\description{
\code{rgamma_vec()} allows vector inputs for \code{shape} and \code{rate}. Unlike
\code{\link[stats:rgamma]{rgamma()}}, it always returns a \code{tibble}, not a
vector, and does not accept a \code{scale} parameter (if you have a scale
parameter, use \code{1/scale} to get the rate). It also uses the notion of "size",
rather than "length", but this difference is not likely to affect end users.
}
\seealso{
\code{\link[stats:rgamma]{rgamma()}}
}
\keyword{internal}
