% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_boosted_rt.R
\name{sample_rt_error}
\alias{sample_rt_error}
\title{Bootstrap Rt Estimates Using Past Errors}
\usage{
sample_rt_error(.data, n = 10000)
}
\arguments{
\item{.data}{The output of \code{weight_rt_error()}}

\item{n}{The number of samples to take from each horizon}
}
\value{
A \code{tibble} of shape and rate estimates
}
\description{
\code{sample_rt_error()} performs weighted sampling of past errors in the mean
and coefficient of variation in Rt estimates using the weights from
\code{\link[covidModel:weight_rt_error]{weight_rt_error()}}. It converts
these estimates to shape and rate parameters for summary.
}
