% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirv.R
\name{sirv}
\alias{sirv}
\title{Simulate a SIR Model with Vaccination and Variants}
\usage{
sirv(
  start = Sys.Date(),
  end = Sys.Date() + 60L,
  rt = 1,
  r0 = NULL,
  pop = 937166,
  active_cases = 500,
  total_cases = 91350,
  total_vac = 222550,
  vac_per_day = 3500,
  pct_vac_s = 0.8,
  vac_eff = 0.9,
  vac_eff_ek = 0.6,
  pct_var = 0.1/0.85,
  pct_var_uk = 0.85,
  pct_var_ek = 0,
  detect = 1/3
)
}
\arguments{
\item{start}{\code{Date} or coercible to \code{Date}. The start date for simulation.}

\item{end}{\code{Date} or coercible to \code{Date}. The end date for simulation.}

\item{rt}{\code{numeric}.
The observed reproduction number at the start of simulation;
assumed constant other than population and variant effects.}

\item{r0}{\code{numeric}. The baseline reproduction number of the non-variant
strains. If this is supplied, \code{rt} is ignored.}

\item{pop}{\code{numeric}. The size of the population being modeled.}

\item{active_cases}{\code{numeric}.
The observed number of active cases at the start of simulation.}

\item{total_cases}{\code{numeric}.
The observed number of total cases at the start of simulation.}

\item{total_vac}{\code{numeric}.
The number of vaccinated people at the start of simulation.}

\item{vac_per_day}{\code{numeric}.
The number of people vaccinated per day during simulation.}

\item{pct_vac_s}{\code{numeric} between \code{0} and \code{1}.
The fraction of vaccinated people vaccinated who have never had COVID
(i.e. are in the susceptible, or "S", compartment).}

\item{vac_eff}{\code{numeric} between \code{0} and \code{1}.
Efficacy of vaccination and natural immunity for non-resistant variants.}

\item{pct_var}{\code{numeric} between \code{0} and \code{1}.
Percent of active cases that are variants of some form.}

\item{pct_var_uk}{\code{numeric} between \code{0} and \code{1}.
Percent of variants that are B.1.1.7 (UK) variant or similarly highly
transmissible. Assumed to not overlap with \code{pct_var_ek}.}

\item{pct_var_ek}{\code{numeric} between \code{0} and \code{1}.
Percent of variants that possess the E484K mutation (vaccine resistance).
Assumed to not overlap with \code{pct_var_uk}.}

\item{detect}{\code{numeric} between \code{0} and \code{1}.
The percent of true cases detected; applies to both past and future cases.}

\item{vacc_eff_ek}{\code{numeric} between \code{0} and \code{1}.
Efficacy of vaccination and natural immunity for resistant variants
(with the E484K mutation).}
}
\value{
A \code{tibble} with columns
\verb{time <date>}, \verb{S <dbl>}, \verb{I <dbl>}, \verb{R <dbl>},
\verb{pct_uk <dbl>}, \verb{pct_ek <dbl>}
}
\description{
\code{sirv()} simulates a SIR model with vaccinations, a variant type with
increased transmissibility, and a variant type with immunity resistance. For
simplicity, vaccine- and naturally-acquired immunity are considered
equivalent.
}
