% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirv.R
\name{sirv_fn}
\alias{sirv_fn}
\title{Workhorse Function for \code{sirv()}}
\usage{
sirv_fn(t, y, params, ...)
}
\arguments{
\item{t}{\code{numeric}. The current timestep.}

\item{y}{\code{numeric}. A named vector of current states.}

\item{params}{\code{numeric}. A named vector of input parameters.}

\item{...}{Additional arguments; currently unused.}
}
\value{
A \code{list} with containing a \code{numeric} vector of output derivatives for
the state components
}
\description{
\code{sirv_fn()} updates derivatives for the \code{\link[covidModel:sirv]{sirv()}}
function. It is passed to the \code{func} parameter of
\code{\link[deSolve:ode]{ode()}}.
}
\keyword{internal}
