% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_boosted_rt.R
\name{summarize_rt_error}
\alias{summarize_rt_error}
\title{Compute Updated Summary of Rt Smooth}
\usage{
summarize_rt_error(.data, .ref)
}
\arguments{
\item{.data}{The output of \code{sample_rt_error()}}

\item{.ref}{Reference Rt data}
}
\value{
A \code{tibble} containing a forecast summary for each future-dependent
data point, as given by \code{\link[covidModel:tidy_rt]{tidy_rt()}}
}
\description{
\code{summarize_rt_error()} computes Rt estimates based on the bootstrapping
performed in \code{sample_rt_error()}.
}
\seealso{
Other internal: 
\code{\link{bind_boosted_rt}()},
\code{\link{boost_rt}()},
\code{\link{cv_rt}()},
\code{\link{reduce_rt_error}()},
\code{\link{validate_rt}()},
\code{\link{weight_rt_error}()}
}
\concept{internal}
