% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_unboosted_rt.R
\name{tidy_rt}
\alias{tidy_rt}
\title{Tidy an \code{estimate_R} Object}
\usage{
tidy_rt(rt)
}
\arguments{
\item{rt}{An \code{estimate_R} object}
}
\value{
A \code{covidmodel_rt} object
}
\description{
\code{tidy_rt()} converts an \code{estimate_R} object from the EpiEstim package to the
\code{tibble} subclass \code{covidmodel_rt} with a \code{serial_interval} attribute. Works
when \code{estimate_R()} is called with \code{method = "parametric_si"}.
}
