% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_decomposition.R
\name{validate_decomposition}
\alias{validate_decomposition}
\title{Map a Dataset to an STL Decomposition Using Data Up to a Specific Time Point}
\usage{
validate_decomposition(.i, .data, .ref, period, trend)
}
\arguments{
\item{.i}{The index of the sample; all data up to this point will be used}

\item{.data}{The data to sample}

\item{.ref}{A reference decomposition used to calculate errors}

\item{period}{The length of time to use in seasonal decomposition; can be a
time-based definition (e.g. "1 week") or an integer number of days. If
\code{NULL} or \code{"auto"}, \code{period} is set automatically using the tunable
heuristics in the timetk package.}

\item{trend}{The length of time to use in trend decomposition; can be a
time-based definition (e.g. "1 month") or an integer number of days. If
\code{NULL} or \code{"auto"}, \code{trend} is set automatically using the tunable
heuristics in the timetk package.}
}
\value{
A \code{tibble} containing the time index of the data, as well as columns
\code{sample} (an indicator of the sample to which the data belongs), \code{horizon},
\code{observed}, \code{season}, \code{trend}, \code{remainder}, \code{season_error}, \code{trend_error},
and \code{remainder_error}
}
\description{
\code{validate_decomposition()} creates a sample decomposition and errors from a
dataset, a reference decomposition, and decomposition parameters. It is
designed to be used as a mapping function passed to
\code{\link[purrr:map]{map()}} within
\code{\link[covidModel:cv_decomposition]{cv_decomposition()}}.
}
\seealso{
\code{
  \link[covidModel:cv_decomposition]{cv_decomposition()}
  }
}
