% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_boosted_rt.R
\name{validate_rt}
\alias{validate_rt}
\title{Compute Errors for Rt Smoothing}
\usage{
validate_rt(
  .data,
  .ref,
  incid,
  .t,
  serial_interval_mean,
  serial_interval_sd,
  half_trend
)
}
\arguments{
\item{.data}{The cross-validation results from \code{cv_linelist_decomposition()}}

\item{.ref}{Reference Rt}

\item{incid}{The quoted name of a numeric column containing the incidence
curve}

\item{.t}{The quoted name of a date column corresponding to the observations
in \code{incid}}

\item{serial_interval_mean}{The average number of days between infection of a
primary case and a secondary case}

\item{serial_interval_sd}{The standard deviation of the number of days
between infection of a primary case and a secondary case}

\item{half_trend}{The portion of the trend conditional on future data}
}
\value{
A tibble of rt estimates and associated errors in each summary
statistic
}
\description{
\code{validate_rt()} calculates Rt and "forecast" errors from a
a sample decomposition, a reference decomposition, and decomposition
parameters. It is designed to be used as a mapping function passed to
\code{\link[purrr:map]{map()}} within
\code{\link[covidModel:cv_decomposition]{cv_decomposition()}}.
}
\seealso{
Other internal: 
\code{\link{bind_boosted_rt}()},
\code{\link{boost_rt}()},
\code{\link{cv_rt}()},
\code{\link{reduce_rt_error}()},
\code{\link{summarize_rt_error}()},
\code{\link{weight_rt_error}()}
}
\concept{internal}
