% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_boosted_rt.R
\name{weight_rt_error}
\alias{weight_rt_error}
\title{Calculate Geometrically Decaying Weights for Rt Sampling}
\usage{
weight_rt_error(.data)
}
\arguments{
\item{.data}{The result of \code{reduce_rt_error()}}
}
\value{
The input with a \code{wt} column containing sampling weights
}
\description{
\code{weight_rt_error()} assigns weights according to a geometric decay curve
parameterized by the length of time from the current timepoint and the
first-order autoregressive coefficient, which is a measure of geometric
dependence in the absence of other modeling terms. This function is used
internally by \code{\link[covidModel:boost_rt]{boost_rt()}}.
}
\seealso{
Other internal: 
\code{\link{bind_boosted_rt}()},
\code{\link{boost_rt}()},
\code{\link{cv_rt}()},
\code{\link{reduce_rt_error}()},
\code{\link{summarize_rt_error}()},
\code{\link{validate_rt}()}
}
\concept{internal}
