% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts_fit.R
\name{bsts_fit}
\alias{bsts_fit}
\title{Bayesian Structural Time Series: Fit Model}
\usage{
bsts_fit(state, .data = state[[".data"]], iterations = 10000, ...)
}
\arguments{
\item{state}{A list specifying the state sub-models}

\item{.data}{The data to use for estimation}

\item{iterations}{The number of MCMC iterations to perform}

\item{...}{Futher arguments to pass to
\code{\link[bsts:bsts]{bsts()}}}
}
\description{
Fit a bsts model.
}
