% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts_trend.R
\name{bsts_trend}
\alias{bsts_trend}
\title{Bayesian Structural Time Series: Trend Components}
\usage{
bsts_trend(
  state = list(),
  .data = state[[".data"]],
  method = c("semilocal", "local", "robust", "level"),
  ...
)
}
\arguments{
\item{state}{A list of state components you wish to add to. If omitted,
an empty list will be assumed. This argument is named \code{state.specification}
in bsts.}

\item{method}{Which trend model to use. Choose from \code{"semilocal"}
(the default), \code{"local"}, \code{"robust"}, or \code{"level"}.}

\item{...}{Additional arguments to pass to methods; see the methods above
for details}

\item{y}{The time series to be modeled, as a numeric vector. Unlike bsts,
this is piped forward as part of the state if defined, so you only need
to specify it once (at the beginning of the model-building pipeline).}
}
\value{
A list with the elements necessary to specify the chosen trend model
}
\description{
\code{bsts_trend()} is a generic that wraps the various trend models in the
bsts package into a user-friendly interface. Model-specific arguments are
passed via \code{...}; see the four methods (
\code{\link[bsts_trend.semilocal]{"semilocal"}},
\code{\link[bsts_trend.local]{"local"}},
\code{\link[bsts_trend.robust]{"robust"}}, and
\code{\link[bsts_trend.level]{"level"}}
) for details of those arguments.
}
\seealso{
\code{\link[bsts:bsts]{bsts()}},
\code{\list[bsts:AddSemilocalLinearTrend]{AddSemilocalLinearTrend()}},
\code{\list[bsts:AddLocalLinearTrend]{AddLocalLinearTrend()}},
\code{\list[bsts:AddStudentLocalLinearTrend]{AddStudentLocalLinearTrend()}},
\code{\list[bsts:AddLocalLevel]{AddLocalLevel()}}
}
