% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{functionalize}
\alias{functionalize}
\title{Integrate a Vector w.r.t. Time
\code{integrate_dt()}}
\usage{
functionalize(.data, .x, .t = NULL, slices = 1000L)
}
\description{
Integrate a Vector w.r.t. Time
\code{integrate_dt()}
}
